sess/driver: allow flags more than once
Rust uses getopts for some argument parsing, which allows specifying
whether a flag is accepted once or multiple times, using `Opt` vs `Multi`
and `Flag` vs `FlagMulti` (depending on whether a flag takes a value).
However, getopts expects this to be used only for flags where every
use of the flag is useful, not when the intention is for a repeated
use of a flag to override previous uses. It provides no function for
accessing a single use of a flag that isn't the first use of a flag.

This commit changes flags from `Opt` to `Multi` and from `Flag` to
`FlagMulti`, and introduces a "last use" variant of the getopt functions
using an extension trait, and then switches to using those throughout
the compiler.

This approach unfortunately means that previous
uses aren't checked at all, only the last use. For example,
`--edition=foo --edition=2024` would be accepted.