libs: relax `MetaSized` bounds
core uses extern type for the `VTable` type as part of the `DynMetadata`
type. It is used as part of the `NonNull` pointer type, which previously
had a `?Sized` bound, which is now a `MetaSized` bound, which the
`VTable` does not satisfy. `NonNull`'s bounds need relaxed before
`?Sized` can begin to be migrated to `MetaSized`. This is somewhat viral
and ends up with lots of relaxations throughout `core`.