hir_analysis: add `MetaSized` bounds with `?Sized`
Opting-out of `Sized` with `?Sized` is now equivalent to adding a
`MetaSized` bound, and opting-out of `MetaSized` with `?MetaSized` is
equivalent to having no bounds whatsoever - this commit implements this
change in `rustc_hir_analysis::hir_ty_lowering`.

`MetaSized` is also added as a supertrait of all traits, as this is
necessary to preserve backwards compatibility.

It's possible to build core/alloc/std with these changes but not all
of the tests pass yet.