trait_sel: re-order host effect evaluations
When evaluating a host effect predicate on an opaque in `addr2line`,
running `evaluate_host_effect_from_bounds` first would create an opaque
type which wouldn't be removed from opaque type storage and that would
cause a delayed bug - creating a small reproduction of this has been
tricky but if `builtin_impls` checking is moved below
`evaluate_host_effect_from_item_bounds` then it will trigger.

The correct fix is probably to work out why the opaque type is created
and then not used, but this was particularly tricky. Given that this root
cause wasn't identified, writing a test case proved difficult.