trait_sel: normalise obligations from builtin impls
Obligations produced by builtin impls need to be normalised so as not to
later be compared to normalised candidates.

For example, during the compilation of std, there will be a
`thread::Inner as const MetaSized` obligation, the builtin impl will
create a `<u32 as AtomicPrimitive>::AtomicInner as const MetaSized`
obligation, which if unnormalised, will end up in
`evaluate_host_effect_from_item_bounds` and then `match_candidate` being
compared with a normalised candidate and failing.