trait_sel: temporary impl `Sized` for scalable vecs
Scalable vectors should behave like other value types - being permitted
as arguments, locals, return types, etc. This requires that these types
be `Sized`, despite not meeting the definition and requirements of
`Sized`. `feature(sized_hierarchy)` is being implemented to enable
scalable vectors to be non-`const Sized`, but in the mean time, add
a builtin impl for scalable vectors. This must be addressed prior to
stabilisation.